refactor: 예약 JDBC Repository를 JPA Repository로 변경

+++ src/main/java/roomescape/service/ReservationService.java
@@ -2,7 +2,7 @@ package roomescape.service;
 
 import java.time.LocalDate;
 import java.util.List;
-import org.springframework.dao.DuplicateKeyException;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import roomescape.domain.Member;
 import roomescape.domain.Reservation;
@@ -12,6 +12,7 @@ import roomescape.dto.reservation.UserReservationRequest;
 import roomescape.exception.DuplicateContentException;
 import roomescape.exception.NotFoundException;
 import roomescape.repository.ReservationRepository;
+import roomescape.repository.ReservationSpecification;
 
 @Service
 public class ReservationService {
@@ -35,13 +36,11 @@ public class ReservationService {
     }
 
     private ReservationResponse createReservation(Reservation reservation) {
-        try {
-            long id = reservationRepository.save(reservation);
-            Reservation newReservation = new Reservation(id, reservation.getMember(), reservation.getDate(), reservation.getTime(), reservation.getTheme());
-            return ReservationResponse.from(newReservation, newReservation.getTime(), newReservation.getTheme());
-        } catch (DuplicateKeyException e) {
+        if (reservationRepository.existsByDateAndTimeIdAndThemeId(reservation.getDate(), reservation.getTime().getId(), reservation.getTheme().getId())) {
             throw new DuplicateContentException("[ERROR] 해당 날짜와 테마로 이미 예약된 내역이 존재합니다.");
         }
+        Reservation newReservation = reservationRepository.save(reservation);
+        return ReservationResponse.from(newReservation, newReservation.getTime(), newReservation.getTheme());
     }
 
     public List<ReservationResponse> findAllReservationResponses() {
@@ -53,7 +52,8 @@ public class ReservationService {
     }
 
     public List<ReservationResponse> searchReservations(Long themeId, Long memberId, LocalDate from, LocalDate to) {
-        List<Reservation> searchResults = reservationRepository.findByThemeMemberDateRange(themeId, memberId, from, to);
+        Specification<Reservation> specification = ReservationSpecification.getReservationSpecification(themeId, memberId, from, to);
+        List<Reservation> searchResults = reservationRepository.findAll(specification);
 
         return searchResults.stream()
                 .map(reservation -> ReservationResponse.from(reservation, reservation.getTime(), reservation.getTheme()))
@@ -61,10 +61,10 @@ public class ReservationService {
     }
 
     public void deleteReservation(Long id) {
-        int deletedReservationCount = reservationRepository.deleteById(id);
-
-        if (deletedReservationCount == 0) {
-            throw new NotFoundException("[ERROR] 등록된 예약번호만 삭제할 수 있습니다. 입력된 번호는 " + id + "입니다.");
+        if (reservationRepository.findById(id).isEmpty()) {
+            throw new NotFoundException("[ERROR] 등록된 예약만 삭제할 수 있습니다. 입력된 번호는 " + id + "입니다.");
         }
+
+        reservationRepository.deleteById(id);
     }
 }
